fix copilot changes
