+++ frontend/src/components/dashboard/BarChartComponent.tsx
@@ -13,7 +13,6 @@ import {
 } from "recharts";
 import type { ChartData } from "@/types/chart";
 import ChartCard from "./ChartCard";
-import type { GenericDataPoint } from "@/lib/chartUtils";
 
 interface Props {
   title: string;
@@ -24,49 +23,14 @@ interface Props {
   yDomain?: [number, number] | ["auto", "auto"];
 }
 
-// --- helper for outlier-aware Y domain ---
-function computeYDomain(
-  points: GenericDataPoint[],
-  yKeys: string[]
-): [number, number] | ["auto", "auto"] {
-  if (!points.length || !yKeys.length) return ["auto", "auto"];
-
-  const values = points.flatMap((p) =>
-    yKeys.map((key) => Number(p[key])).filter((v) => !isNaN(v))
-  );
-
-  if (!values.length) return ["auto", "auto"];
-
-  values.sort((a, b) => a - b);
-  const q1 = values[Math.floor(values.length * 0.25)];
-  const q3 = values[Math.floor(values.length * 0.75)];
-  const iqr = q3 - q1;
-  const upperFence = q3 + 1.5 * iqr;
-
-  const nonOutlierMax = Math.max(...values.filter((v) => v <= upperFence));
-  const safeMax = Math.max(nonOutlierMax, 1);
-
-  return [0, safeMax];
-}
-
 export default function BarChart({
   title,
   subtitle,
   data,
   isLoading,
   error,
-  yDomain,
+  yDomain = ["auto", "auto"],
 }: Props) {
-  // Compute yDomain dynamically if not passed
-  const finalYDomain =
-    yDomain ||
-    (data
-      ? computeYDomain(
-          data.dataPoints as unknown as GenericDataPoint[],
-          data.yAxisKeys.map((y) => y.key)
-        )
-      : ["auto", "auto"]);
-
   return (
     <ChartCard
       title={title}
@@ -93,7 +57,7 @@ export default function BarChart({
               tick={{ fill: "var(--foreground)" }}
             />
             <YAxis
-              domain={finalYDomain} // ✅ dynamic Y-axis scaling
+              domain={yDomain}
               unit={` ${data.yAxisUnit}`}
               fontSize={12}
               width={60}
